added generic filter parameters moved generic models into generic folder updated imports added new event search fields
