+++ Singer.API/ClientApp/src/app/modules/admin/components/singerevents/singerevent-details/singerevent-details.component.css
@@ -1,5 +1,5 @@
 /* Form layout */
-.details-form-container {
+.form {
    display: flex;
    flex-direction: column;
    justify-content: flex-start;
