+++ Singer.API/ClientApp/src/app/modules/core/models/generics/generic-filter-parameters.model.ts
@@ -1,0 +1,3 @@
+export class GenericFilterParameters {
+   name: string;
+}
