+++ Singer.API/ClientApp/src/app/modules/core/models/generics/generic-model.ts
