+++ Singer.API/ClientApp/src/app/modules/core/services/generic-service.ts
@@ -1,4 +1,4 @@
-import { GenericModel } from '../models/generic-model';
+import { GenericModel } from '../models/generics/generic-model';
 import {
    HttpClient,
    HttpParams,
