+++ Singer.API/ClientApp/src/app/modules/dashboard/components/event-search/event-search.component.html
@@ -1,7 +1,7 @@
 <div class="container">
-   <form [formGroup]="formControlGroup" class="details-form-container">
-      <!-- Name Fields -->
-      <!-- Lastname Field -->
+   <form [formGroup]="formControlGroup" class="form">
+
+      <!-- StartDate Field -->
       <mat-form-field>
          <input matInput [min]="currentDate" [matDatepicker]="startDatePicker" placeholder="Startdatum"
             formControlName="startDateControl" (click)="startDatePicker.open()" />
@@ -9,9 +9,9 @@
          </mat-datepicker-toggle>
          <mat-datepicker #startDatePicker touchUi [startAt]="currentDate" disabled="false">
          </mat-datepicker>
-
       </mat-form-field>
 
+      <!-- EndDate Field -->
       <mat-form-field>
          <input matInput [min]="currentDate" [matDatepicker]="endDatePicker" placeholder="Einddatum"
             formControlName="endDateControl" (click)="endDatePicker.open()" />
@@ -19,8 +19,9 @@
          </mat-datepicker-toggle>
          <mat-datepicker #endDatePicker touchUi [startAt]="currentDate" disabled="false">
          </mat-datepicker>
-
       </mat-form-field>
+
+      <!-- Location Field -->
       <mat-form-field>
          <mat-select formControlName="locationControl" placeholder="Locatie">
             <mat-option *ngFor="let availableLocation of availableLocations" [value]="availableLocation">
@@ -28,8 +29,40 @@
             </mat-option>
          </mat-select>
       </mat-form-field>
-      <button type="button" mat-raised-button color="primary" (click)="submitForm()">
+
+      <!-- AgeGroup Field -->
+      <mat-form-field>
+         <mat-select formControlName="locationControl" placeholder="Locatie">
+            <mat-option *ngFor="let availableLocation of availableLocations" [value]="availableLocation">
+               {{ availableLocation.name }}
+            </mat-option>
+         </mat-select>
+      </mat-form-field>
+
+      <!-- Name Field -->
+      <mat-form-field>
+         <mat-select formControlName="locationControl" placeholder="Locatie">
+            <mat-option *ngFor="let availableLocation of availableLocations" [value]="availableLocation">
+               {{ availableLocation.name }}
+            </mat-option>
+         </mat-select>
+      </mat-form-field>
+
+      <!-- Price Field -->
+      <mat-form-field>
+         <mat-select formControlName="locationControl" placeholder="Locatie">
+            <mat-option *ngFor="let availableLocation of availableLocations" [value]="availableLocation">
+               {{ availableLocation.name }}
+            </mat-option>
+         </mat-select>
+      </mat-form-field>
+
+      <!-- Buttons -->
+      <button type="button" mat-stroked-button color="primary" (click)="setFilter()">
          Zoek
+      </button>
+      <button type="button" mat-strokedstroked-button (click)="resetFilter()">
+         Reset Filter
       </button>
    </form>
 </div>
